+++ lib/test/clojure_lsp/shared_test.clj
@@ -1,5 +1,6 @@
 (ns clojure-lsp.shared-test
   (:require
+   [babashka.fs :as fs]
    [clojure-lsp.shared :as shared]
    [clojure-lsp.test-helper.internal :as h]
    [clojure.test :refer [are deftest is testing]]
@@ -280,19 +281,15 @@
         "file:///c:/c.clj"))))
 
 (deftest dir-uris->file-uris-test
-  (swap! (h/db*) shared/deep-merge {:settings {:source-paths #{(h/file-uri "file:///user/project/src")
-                                                               (h/file-uri "file:///user/project/test")}}
-                                    :project-root-uri (h/file-uri "file:///user/project")})
-  (h/load-code (h/code "(ns foo.bar)") (h/file-uri "file:///user/project/src/foo/bar.clj"))
-  (h/load-code (h/code "(ns foo.baz)") (h/file-uri "file:///user/project/src/foo/baz.clj"))
-  (h/load-code (h/code "(ns bar.foo)") (h/file-uri "file:///user/project/test/bar/foo.clj"))
-  (h/load-code (h/code "(ns bar.baz)") (h/file-uri "file:///user/project/other/bar/baz.clj"))
   (testing "when the dir-uri is a dir inside source-path"
-    (is (= [(h/file-uri "file:///user/project/src/foo/bar.clj")
-            (h/file-uri "file:///user/project/src/foo/baz.clj")]
-           (shared/dir-uris->file-uris [(h/file-uri "file:///user/project/src")]
-                                       (h/db)))))
+    (with-redefs [fs/glob (constantly [(h/file-uri "file:///user/project/src/foo/bar.clj")
+                                       (h/file-uri "file:///user/project/src/foo/baz.clj")])
+                  fs/canonicalize identity]
+      (is (= [(h/file-uri "file:///user/project/src/foo/bar.clj")
+              (h/file-uri "file:///user/project/src/foo/baz.clj")]
+             (shared/dir-uris->file-uris [(h/file-uri "file:///user/project/src")])))))
   (testing "when the dir-uri is absolute file URI"
-    (is (= [(h/file-uri "file:///user/project/src/foo/bar.clj")]
-           (shared/dir-uris->file-uris [(h/file-uri "file:///user/project/src/foo/bar.clj")]
-                                       (h/db))))))
+    (with-redefs [fs/glob (constantly [(h/file-uri "file:///user/project/src/foo/bar.clj")])
+                  fs/canonicalize identity]
+      (is (= [(h/file-uri "file:///user/project/src/foo/bar.clj")]
+             (shared/dir-uris->file-uris [(h/file-uri "file:///user/project/src/foo/bar.clj")]))))))